Fix handle leak in parent process
Since the parent process usually never logs, this module will hold onto the file handle forever, which somewhat defeats the point of log rotation.

To fix this, we for the parent process, we close the handle immediately after opening it, and add support to ap_lock_log for reopening it if we ever log.